% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessPath.r
\name{accessPath}
\alias{accessPath}
\title{Path Accessibility}
\usage{
accessPath(
  wholenet,
  subbox = "NA",
  more = FALSE,
  fromID = NULL,
  toID = NULL,
  status = NULL
)
}
\arguments{
\item{wholenet}{Reference network with node IDs and lon-lat coordinates.}

\item{subbox}{Sub-region of the reference network. The case subbox = "NA" is prepared
for the shortest path.}

\item{more}{If TRUE, will return more info including the distance "d", the weighted
distance "d_weighted", the "time", and the weighted time "time_weighted".}

\item{fromID}{The starting node ID as expected.}

\item{toID}{The ending node ID as expected.}

\item{status}{Two possible status for this function. If status = NULL, as the
default, a full list will be returned. If status = "ShortPath", only the
shortest path will be returned.}
}
\value{
The output is a list of a new network with the created nodes
"addedNet", a boundary-crossed segment "crossSeg", and a segemnt with the
added new node "newCrossSeg". It also includes the shortest path "shortPath" and
the coordinates of the created boundary nodes "xyFrame". If the status is
"ShortPath", only the thortest path will be returned.
}
\description{
This function creates new boundary nodes for a sub-region using the method of
detecting the shortest path between two nodes.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
regn <- c(-1.24, 51.71, -1.19, 51.79) # Selected a sub-region
subbox <- vec2bbox(regn)
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.od.nodes <- ODlist[[1]]
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)
oxford.od.matrix <- ODwithdemand(oxford.od)
OxfordOD <- renameODdata(oxford.od.matrix)
odlist2 <- od2network(OxfordOD, oxford.od.nodes, cntrnet)
OxfordODre <- odlist2[[1]]

in2outFrom <- tailor(subbox, OxfordODre, od="from")
in2outTo <- tailor(subbox, OxfordODre, od="to", "outside")
fromIDs <- getFromToID(in2outFrom, "from")
toIDs <- getFromToID(in2outTo, "to")

listp <- accessPath(cntrnet, subbox, more=FALSE,
                    fromID = fromIDs[4], toID = toIDs[5])
shortPath <- listp$shortPath
xyFrame <- listp$xyFrame
newCrossSeg <- listp$newCrossSeg

shownetwork(vec2bbox(location), wholenet, cntrnet)
addbox(regn, "blue")
addSegments(shortPath, "red", lwd = 2, showEnds = TRUE)
with(
  xyFrame,
  points(x, y, pch = 4, col= "green")
)
addSegments(newCrossSeg, "yellow", showEnds = FALSE)

}
\keyword{accessPath}
