% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addbox.r
\name{addbox}
\alias{addbox}
\title{Show a Boundary-Box}
\usage{
addbox(location, col = NULL)
}
\arguments{
\item{location}{A vector contains coordinates of min-x, min-y, max-x and max-y.}

\item{col}{The specified colour.}
}
\value{
The output is a box added to the plot.
}
\description{
This function adds a box to the current plot.
}
\examples{
wholenet <- getnet("oxford uk")
bbox <- osmdata::getbb("oxford uk")
shownetwork(bbox, wholenet)
addbox(c(bbox), col = "blue")

}
\keyword{addbox}
