% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcapacity.r
\name{addcapacity}
\alias{addcapacity}
\title{Add Capacities to a Network}
\usage{
addcapacity(linksdata, specified = NULL)
}
\arguments{
\item{linksdata}{A network with the "highway" column factorized by "trunk",
"primary", "secondary", "tertiary", "trunk_link", "primary_link",
"secondary_link" and "tertiary_link".}

\item{specified}{Prepared for further development on specified numerical
capacities.}
}
\value{
The output is a network data frame with the added capacities.
}
\description{
This function adds numerical capacities to the network data frame.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.main.nodes <- ODlist[[2]]
oxford.links.df <- links(oxford.main.nodes, cntrnet)
OxfordLinks <- addcapacity(oxford.links.df)
head(OxfordLinks)

}
\keyword{addcapacity}
