% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossedPnts.r
\name{addCrossedPnts}
\alias{addCrossedPnts}
\title{Get Boundary-Crossed Points}
\usage{
addCrossedPnts(crossSeg, regn, more = FALSE)
}
\arguments{
\item{crossSeg}{A data frame of boundary-crossed segments.}

\item{regn}{The boundary of a region.}

\item{more}{If TRUE, will return more info including the distance "d", the weighted
distance "d_weighted", the "time", and the weighted time "time_weighted".}
}
\value{
The output is a list of a new segments data frame with added boundary
nodes "crossedAdded" and a data frame of the coordinates of the created
boundary nodes "xyFrame".
}
\description{
This function gets the coordinates of boundary-crossed points for the
provided boundary-crossed segments.
}
\examples{
wholenet <- getnet("oxford uk")
crossSeg <- wholenet[1, ]

x1 <- mean(c(crossSeg$from_lon, crossSeg$to_lon))
y1 <- mean(c(crossSeg$from_lat, crossSeg$to_lat))
x2 <- max(c(crossSeg$from_lon, crossSeg$to_lon)) + abs(x1)
y2 <- max(c(crossSeg$from_lat, crossSeg$to_lat)) + abs(y1)
regn <- c(x1, y1, x2, y2)
addCrossedPnts(crossSeg, regn, more = TRUE)$
  crossedAdded[, c("d", "d_weighted", "time", "time_weighted")]
addCrossedPnts(crossSeg, regn, more = FALSE)$
  crossedAdded[, c("d", "d_weighted", "time", "time_weighted")]
addCrossedPnts(crossSeg, regn, more = TRUE)$xyFrame

}
\keyword{addCrossedPnts}
