% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addODnodeID.r
\name{addODnodeID}
\alias{addODnodeID}
\title{Add Node IDs}
\usage{
addODnodeID(mainnodes, ODdata)
}
\arguments{
\item{mainnodes}{Provided node IDs.}

\item{ODdata}{A data frame that needs node IDs.}
}
\value{
The output is a new data frame with node IDs, "from_nodeID" and
"to_nodeID".
}
\description{
This function adds node IDs to the OD data.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
regn <- c(-1.24, 51.71, -1.19, 51.79) # Selected a sub-region
subbox <- vec2bbox(regn)
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.od.nodes <- ODlist[[1]]
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)

c("from_nodeID", "to_nodeID") \%in\% names(oxford.od)

}
\keyword{addODnodeID}
