% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossedPoint.r
\name{crossedPoint}
\alias{crossedPoint}
\title{Crossed-Point of Two Segments}
\usage{
crossedPoint(a1, b1, c1, d1, a2, b2, c2, d2)
}
\arguments{
\item{a1}{The x-coordinate of the first point in segment-1.}

\item{b1}{The y-coordinate of the first point in segment-1.}

\item{c1}{The x-coordinate of the second point in segment-1.}

\item{d1}{The y-coordinate of the second point in segment-1.}

\item{a2}{The x-coordinate of the first point in segment-2.}

\item{b2}{The y-coordinate of the first point in segment-2.}

\item{c2}{The x-coordinate of the second point in segment-2.}

\item{d2}{The y-coordinate of the second point in segment-2.}
}
\value{
The output is the coordinates of the crossed-point. It will be "NULL"
if the segments are parallel.
}
\description{
This function gets the coordinates of the crossed-point of two segments.
}
\examples{
a1=0; b1=0; c1=2; d1=2; a2=0; b2=2; c2=2; d2=0
plot(c(a1,c1), c(b1,d1), xlim=c(-1,3), ylim=c(-1,3), type="l")
segments(a2, b2, c2, d2)
xy <- crossedPoint(a1, b1, c1, d1, a2, b2, c2, d2)
xy
points(xy[1], xy[2], col="red")

a1=0; b1=0; c1=0; d1=2; a2=1; b2=2; c2=1; d2=0
plot(c(a1,c1), c(b1,d1), xlim=c(-1,3), ylim=c(-1,3), type="l")
segments(a2, b2, c2, d2)
xy <- crossedPoint(a1, b1, c1, d1, a2, b2, c2, d2)
is.null(xy)

}
\keyword{crossedPoint}
