% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromToID.r
\name{getFromToID}
\alias{getFromToID}
\title{Get Specified Location IDs}
\usage{
getFromToID(dataframe, status)
}
\arguments{
\item{dataframe}{A data frame with "from_id" and "to_id".}

\item{status}{To specify the "from" or "to".}
}
\value{
The output is the specified IDs.
}
\description{
This function gets location IDs based on the given conditions on "from" or
"to".
}
\examples{
regn <- c(-1.20, 51.71, -1.19, 51.72)
network <- getnet(regn)
fromIDs <- getFromToID(network, "from")
head(fromIDs)

}
\keyword{getFromToID}
