% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnet.r
\name{getnet}
\alias{getnet}
\title{Generate a Network}
\usage{
getnet(location, wt_for = "motorcar")
}
\arguments{
\item{location}{The location of a preferred network.}

\item{wt_for}{The specified profile for the network. The default setting is
"motorcar".}
}
\value{
The output is the network with the specified profile.
}
\description{
This function generates a network according to the given location.
}
\examples{
oxford <- getnet("oxford uk")
shownetwork(osmdata::getbb("oxford uk"), oxford)

smallerNet <- getnet(c(-1.24, 51.71, -1.19, 51.79))
shownetwork(c(-1.24, 51.71, -1.19, 51.79), smallerNet)

}
\keyword{getnet}
