% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getODnodes.r
\name{getODnodes}
\alias{getODnodes}
\title{Set Node IDs for OD Data}
\usage{
getODnodes(network, ODdata)
}
\arguments{
\item{network}{The referred network with node IDs. In general, it is a
contracted network.}

\item{ODdata}{An OD data that needs node IDs.}
}
\value{
The output is a list with a new OD data, where node IDs are added,
and a data frame with main nodes.
}
\description{
This function adds node IDs to the OD data by using the
"match_points_to_graph" function (in the "dodgr" package) and referring to a
given network with node IDs.
}
\examples{
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.od.nodes <- ODlist[[1]]
oxford.main.nodes <- ODlist[[2]]
head(oxford.od.nodes)
head(oxford.main.nodes)

}
\keyword{getODnodes}
