% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSortestPath.r
\name{getSortestPath}
\alias{getSortestPath}
\title{Get Shortest Path}
\usage{
getSortestPath(network, fromID = fromID, toID = toID)
}
\arguments{
\item{network}{A given network. In general, it is a contracted network.}

\item{fromID}{The starting location represented by a from_id.}

\item{toID}{The ending location represented by a to_id.}
}
\value{
The output is a data frame of the shortest path within the given
network.
}
\description{
This function gets the shortest path for the provided from_id and to_id
within a given network.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)

set.seed(1)
fromID <- cntrnet[sample(1:nrow(cntrnet), 1), "from_id"]
set.seed(2)
toID <- cntrnet[sample(1:nrow(cntrnet), 1), "to_id"]
ShortPath <- getSortestPath(cntrnet, fromID = fromID, toID = toID)

shownetwork(vec2bbox(location), wholenet, cntrnet)
addSegments(ShortPath, "red", lwd = 2, showEnds = TRUE)

}
\keyword{getSortestPath}
