% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{links}
\alias{links}
\title{Link to Node IDs}
\usage{
links(mainnodes, network)
}
\arguments{
\item{mainnodes}{A data frame with main node IDs.}

\item{network}{A network with "from_id" and "to_id." In general, it is a
ontracted network.}
}
\value{
The out put is a network data frame with added node IDs.
}
\description{
This function links the "from_id" and "to_id" to node IDs based on the given two
data frames.
}
\examples{
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.main.nodes <- ODlist[[2]]
oxford.links.df <- links(oxford.main.nodes, cntrnet)
head(oxford.links.df)

}
\keyword{links}
