% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainnet.r
\name{mainnet}
\alias{mainnet}
\title{Get the Main Network}
\usage{
mainnet(network, mainset = NULL)
}
\arguments{
\item{network}{A given network with the road type of "highway".}

\item{mainset}{A provided vector of road type, e.g. c("trunk", "trunk_link").
If NULL, will use the default setting, c("trunk", "trunk_link", "primary",
"primary_link", "secondary", "secondary_link", "tertiary", "tertiary_link").}
}
\value{
The output is a main network with the preferred road type.
}
\description{
This function extracts the main network from a given network based on the
provided road types.
}
\examples{
bbox <- osmdata::getbb("oxford uk")
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
shownetwork(bbox, wholenet, mainnet)

}
\keyword{mainnet}
