% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNodes.r
\name{mergeNodes}
\alias{mergeNodes}
\title{Merge Neighbour Nodes}
\usage{
mergeNodes(network, threshold = NULL)
}
\arguments{
\item{network}{A network with nodes and their distances.}

\item{threshold}{If NULL, will merge nodes based on the mean of the distances.
Otherwise, will merge nodes with the given threshold (as a quantile).}
}
\value{
The output is a list of a data frame with shorter-routes (shorters),
a data frame without shorter-routes (NOshorters), a data frame with repeated
merged-centres (tailorNR), and the merged network without shorter-routes
(merged).
}
\description{
This function merges some close nodes into one single node.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
shownetwork(vec2bbox(location), wholenet, cntrnet)
addbox(location, "blue")
listAll <- mergeNodes(cntrnet, threshold = 0.98)
shorters <- listAll$shorters
merged <- listAll$merged
with(shorters, segments(from_lon, from_lat, to_lon, to_lat, col = "green"))
with(merged, segments(from_lon, from_lat, to_lon, to_lat, col = "red"))

}
\keyword{mergeNodes}
