% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od2network.r
\name{od2network}
\alias{od2network}
\title{Add Coordinates to OD Data}
\usage{
od2network(odData, odNode, wholenet)
}
\arguments{
\item{odData}{The OD data frame, which needs lon-lat coordinates.}

\item{odNode}{A data frame with nodes info.}

\item{wholenet}{The reference network with lon-lat coordinates.}
}
\value{
The output is a list of an OD data frame with lon-lat coordinates and
a data frame contains nodes out of the network.
}
\description{
This function adds lon-lat coorniates to the OD data according to the given
OD nodes.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.od.nodes <- ODlist[[1]]
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)
oxford.od.matrix <- ODwithdemand(oxford.od)
OxfordOD <- renameODdata(oxford.od.matrix)
odlist2 <- od2network(OxfordOD, oxford.od.nodes, cntrnet)
OxfordODre <- odlist2[[1]]
outNet <- odlist2[[2]]

head(OxfordODre)
head(outNet)

}
\keyword{od2network}
