% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODdata.r
\name{ODdata}
\alias{ODdata}
\title{Get the OD Data}
\usage{
ODdata(location)
}
\arguments{
\item{location}{The location where needs the OD data.}
}
\value{
The output is the OD data of the location.
}
\description{
This function generates the OD data for a given location.
}
\examples{
oxford.od <- ODdata("oxford uk")
head(oxford.od)

}
\keyword{ODdata}
