% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODwithdemand.r
\name{ODwithdemand}
\alias{ODwithdemand}
\title{Add Demands to OD Data}
\usage{
ODwithdemand(ODdata)
}
\arguments{
\item{ODdata}{The OD data with "car_driver" counts.}
}
\value{
The output is an OD data with demands generated by using the
"car_driver" counts.
}
\description{
This function adds demands to the OD data.
}
\examples{
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)
oxford.od.matrix <- ODwithdemand(oxford.od)
head(oxford.od.matrix)

}
\keyword{ODwithdemand}
