% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameODdata.r
\name{renameODdata}
\alias{renameODdata}
\title{Rename the OD Data}
\usage{
renameODdata(ODdata)
}
\arguments{
\item{ODdata}{The original OD data.}
}
\value{
The output is the OD data with names, "demand", "fromNode"
and "toNode".
}
\description{
This function renames the OD data by using column names: "demand", "fromNode"
and "toNode".
}
\examples{
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)
oxford.od.matrix <- ODwithdemand(oxford.od)
OxfordOD <- renameODdata(oxford.od.matrix)
head(OxfordOD)

}
\keyword{renameODdata}
