% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetIDs.r
\name{resetIDs}
\alias{resetIDs}
\title{Reset Node IDs}
\usage{
resetIDs(network)
}
\arguments{
\item{network}{A network data frame with node IDs, location IDs and the
lon-lat coordinates.}
}
\value{
The output is a network with the reset node IDs.
}
\description{
This function resets node IDs with increasing positive integers.
}
\examples{
wholenet <- getnet("oxford uk")
oxford.od <- ODdata("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
ODlist <- getODnodes(cntrnet, oxford.od)
oxford.od.nodes <- ODlist[[1]]
oxford.main.nodes <- ODlist[[2]]
oxford.od <- addODnodeID(oxford.main.nodes, oxford.od)
oxford.od.matrix <- ODwithdemand(oxford.od)
OxfordOD <- renameODdata(oxford.od.matrix)
odlist2 <- od2network(OxfordOD, oxford.od.nodes, cntrnet)
OxfordODre <- odlist2[[1]]
newODdata <- resetIDs(OxfordODre)
head(newODdata)

}
\keyword{resetIDs}
