% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shownetwork.r
\name{shownetwork}
\alias{shownetwork}
\title{Show Networks}
\usage{
shownetwork(
  bbox,
  network,
  contracted = NULL,
  netnodes = NULL,
  tittle = NULL,
  nodeID = TRUE,
  capacity = NULL,
  extended = 0.02
)
}
\arguments{
\item{bbox}{A given region with the range of longitudes and latitudes.}

\item{network}{A network as the background network.}

\item{contracted}{A network on top of the background network. In general, it
is a contracted network.}

\item{netnodes}{The node coordinates. If it is NOT NULL, nodes will be shown
on the networks.}

\item{nodeID}{If TRUE, node IDs will be shown on the network.}

\item{capacity}{If the capacities are given, the routes thickness will change
according to the capacities.}

\item{extended}{A ratio on the extended boundary of the current plot.}
}
\value{
The output is a map of the networks.
}
\description{
This function shows the networks with a given region.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
cntrnet <- contractNet(mainnet)
shownetwork(vec2bbox(location), wholenet, mainnet, tittle = "Main Network On The Background")
shownetwork(vec2bbox(location), wholenet, cntrnet)

}
\keyword{shownetwork}
