% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor.r
\name{tailor}
\alias{tailor}
\title{Tailor the Network}
\usage{
tailor(bbox, network, od = NULL, out = NULL)
}
\arguments{
\item{bbox}{A given boundary box.}

\item{network}{A network that needs to be tailored.}

\item{od}{The OD info on "from" or "to" to clarify going outside the box or
going inside the box respectively.}

\item{out}{If NULL, the inside-to-outside data will be returned.
If "outside", the outside-to-inside data will be returned.}
}
\value{
The output is the tailored network.
}
\description{
This function tailors a network with the given boundary box.
}
\examples{
location <- c(osmdata::getbb("oxford uk"))
wholenet <- getnet("oxford uk")
mainnet <- mainnet(wholenet)
shownetwork(vec2bbox(location), wholenet, tittle = "Whole Network")
regn <- c(-1.24, 51.71, -1.19, 51.79) # Selected a sub-region
subbox <- vec2bbox(regn)
subregion <- tailor(subbox, wholenet)
shownetwork(vec2bbox(location), subregion, tittle = "Tailored Network")

}
\keyword{tailor}
