% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2bbox.r
\name{vec2bbox}
\alias{vec2bbox}
\title{Vector to Boundary Box}
\usage{
vec2bbox(vec)
}
\arguments{
\item{vec}{The vector of the range of longitudes and latitudes in the form
of c(min_lon, min_lat, max_lon, max_lat).}
}
\value{
The output is a boundary box.
}
\description{
This function changes a vector to the boundary box frame.
}
\examples{
osmdata::getbb("oxford uk")
c(osmdata::getbb("oxford uk"))
vec2bbox(c(osmdata::getbb("oxford uk")))

}
\keyword{vec2bbox}
