% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.r
\name{wrapper}
\alias{wrapper}
\title{Wrap Link-Data and OD-Data}
\usage{
wrapper(location, wt_for = "motorcar", mainset = NULL)
}
\arguments{
\item{location}{The location of a preferred network.}

\item{wt_for}{The specified profile for the network. The default setting is
"motorcar".}

\item{mainset}{A provided vector of road type, e.g. c("trunk", "trunk_link").
If NULL, will use the default setting, c("trunk", "trunk_link", "primary",
"primary_link", "secondary", "secondary_link", "tertiary", "tertiary_link").}
}
\value{
The output is a list contains the original link-data and the refined
OD data.
}
\description{
This function wraps the link-data and the OD data into a list according to a
given location.
}
\examples{

location <- "oxford uk"
twoDataFrame <- wrapper(location)
originalLinkData <- twoDataFrame$linksData
refinedODdata <- twoDataFrame$refinedODdata
head(originalLinkData)
head(refinedODdata)

}
\keyword{wrapper}
